--- a/paket.dependencies
+++ b/paket.dependencies
@@ -2,6 +2,7 @@ source https://nuget.org/api/v2
 nuget BenchmarkDotNet
 nuget BenchmarkDotNet.Diagnostics.Windows
 nuget FParsec
+nuget Hopac
 nuget Newtonsoft.Json
 nuget Suave