--- a/src/FSharp.Data.GraphQL/Parser.fs
+++ b/src/FSharp.Data.GraphQL/Parser.fs
@@ -230,7 +230,7 @@ module internal Internal =
     pipe5 (opt(attempt alias)) (token_ws name) (opt(token_ws arguments)) (opt directives) (opt selectionSet)
       (fun oalias name oargs directives oselection ->
          (Field { Alias = oalias; Name = name; Arguments = someOrEmpty oargs;
-                  Directives = someOrEmpty directives; SelectionSet = someOrEmpty oselection }))
+                  Directives = someOrEmpty directives; SelectionSet = match oselection with None -> [] | Some s -> s }))
     <?> "Field"