--- a/src/FSharp.Data.GraphQL/paket.references
+++ b/src/FSharp.Data.GraphQL/paket.references
@@ -1,1 +1,2 @@
 FParsec
+Hopac