--- a/tests/FSharp.Data.GraphQL.Benchmarks/App.config
+++ b/tests/FSharp.Data.GraphQL.Benchmarks/App.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6" />
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
   </startup>
 </configuration>