--- a/tests/FSharp.Data.GraphQL.Benchmarks/paket.references
+++ b/tests/FSharp.Data.GraphQL.Benchmarks/paket.references
@@ -1,2 +1,3 @@
 BenchmarkDotNet
 BenchmarkDotNet.Diagnostics.Windows
+Hopac