--- a/tests/FSharp.Data.GraphQL.Tests/AbstractionTests.fs
+++ b/tests/FSharp.Data.GraphQL.Tests/AbstractionTests.fs
@@ -29,36 +29,36 @@ type Cat =
 type Human = { Name: string; }
 
 type Pet =
-    | Dog of Dog
-    | Cat of Cat
+    | DogCase of Dog
+    | CatCase of Cat
 
 let resolvePet = function
-    | Dog d -> box d
-    | Cat c -> upcast c
+    | DogCase d -> box d
+    | CatCase c -> upcast c
 
 [<Fact>]
 let ``Execute handles execution of abstract types: isTypeOf is used to resolve runtime type for Interface`` () = 
-    let PetType = Define.Interface("Pet", fun () -> [ Define.Field("name", String) ])
+    let PetType = Define.Interface("Pet", fun () -> [| Define.Field("name", String) |])
     let DogType = Define.Object<Dog>(
         name = "Dog", 
         isTypeOf = is<Dog>,
-        interfaces = [ PetType ],
-        fields = [
+        interfaces = [| PetType |],
+        fields = [|
             Define.Field("name", String)
             Define.Field("woofs", Boolean)
-        ])
+        |])
     let CatType = Define.Object<Cat>(
         name = "Cat", 
         isTypeOf = is<Cat>,
-        interfaces = [ PetType ],
-        fields = [
+        interfaces = [| PetType |],
+        fields = [|
             Define.Field("name", String)
             Define.Field("meows", Boolean)
-        ])
+        |])
     let schema = Schema(
-        query = Define.Object("Query", fun () -> [
+        query = Define.Object("Query", fun () -> [|
             Define.Field("pets", ListOf PetType, fun _ _ -> upcast [ { Name = "Odie"; Woofs = true } :> IPet ; upcast { Name = "Garfield"; Meows = false } ])
-        ]), 
+        |]), 
         config = { SchemaConfig.Default with Types = [CatType; DogType] })
     let query = """{
       pets {
@@ -88,22 +88,22 @@ let ``Execute handles execution of abstract types: isTypeOf is used to resolve r
     let DogType = Define.Object<Dog>(
         name = "Dog", 
         isTypeOf = is<Dog>,
-        fields = [
+        fields = [|
             Define.Field("name", String)
             Define.Field("woofs", Boolean)
-        ])
+        |])
     let CatType = Define.Object<Cat>(
         name = "Cat", 
         isTypeOf = is<Cat>,
-        fields = [
+        fields = [|
             Define.Field("name", String)
             Define.Field("meows", Boolean)
-        ])
-    let PetType = Define.Union("Pet", [ DogType; CatType ], resolvePet)
+        |])
+    let PetType = Define.Union("Pet", [| DogType; CatType |], resolvePet)
     let schema = Schema(
-        query = Define.Object("Query", fun () -> [
-            Define.Field("pets", ListOf PetType, fun _ _ -> upcast [ Dog { Name = "Odie"; Woofs = true }; Cat { Name = "Garfield"; Meows = false } ])
-        ]))
+        query = Define.Object("Query", fun () -> [|
+            Define.Field("pets", ListOf PetType, fun _ _ -> upcast [ DogCase { Name = "Odie"; Woofs = true }; CatCase { Name = "Garfield"; Meows = false } ])
+        |]))
     let query = """{
       pets {
         ... on Dog {