--- a/tests/FSharp.Data.GraphQL.Tests/IntrospectionTests.fs
+++ b/tests/FSharp.Data.GraphQL.Tests/IntrospectionTests.fs
@@ -14,8 +14,8 @@ open FSharp.Data.GraphQL.Client
 
 [<Fact>]
 let ``Introspection schema should be serializable back and forth using json`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", String)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", String) |])
     let schema = Schema(root)
     let introResult = schema.AsyncExecute(Introspection.introspectionQuery) |> sync
     let json = Client.Serialization.toJson introResult
@@ -24,8 +24,8 @@ let ``Introspection schema should be serializable back and forth using json`` ()
 
 [<Fact>]
 let ``Core type definitions are considered nullable`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", String)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", String) |])
     let schema = Schema(root)
     let query = """{ __type(name: "String") {
       kind
@@ -54,8 +54,8 @@ let ``Core type definitions are considered nullable`` () =
     
 [<Fact>]
 let ``Default field type definitions are considered non-null`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", String)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", String) |])
     let schema = Schema(root)
     let query = """{ __type(name: "Query") {
       fields {
@@ -96,8 +96,8 @@ let ``Default field type definitions are considered non-null`` () =
     
 [<Fact>]
 let ``Nullabe field type definitions are considered nullable`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", Nullable String)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", Nullable String) |])
     let schema = Schema(root)
     let query = """{ __type(name: "Query") {
       fields {
@@ -135,8 +135,8 @@ let ``Nullabe field type definitions are considered nullable`` () =
     
 [<Fact>]
 let ``Default field args type definitions are considered non-null`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", String, "", [Define.Input("onlyArg", Int)], fun _ () -> null)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", String, "", [| Define.Input("onlyArg", Int) |], fun _ () -> null) |])
     let schema = Schema(root)
     let query = """{ __type(name: "Query") {
       fields {
@@ -181,8 +181,8 @@ let ``Default field args type definitions are considered non-null`` () =
     
 [<Fact>]
 let ``Nullable field args type definitions are considered nullable`` () =
-    let root = Define.Object("Query", [
-        Define.Field("onlyField", String, "", [Define.Input("onlyArg", Nullable Int)], fun _ () -> null)])
+    let root = Define.Object("Query", [|
+        Define.Field("onlyField", String, "", [| Define.Input("onlyArg", Nullable Int) |], fun _ () -> null) |])
     let schema = Schema(root)
     let query = """{ __type(name: "Query") {
       fields {
@@ -224,9 +224,9 @@ let ``Nullable field args type definitions are considered nullable`` () =
 
 [<Fact(Skip = "Investigate and fix the test, introspeciton is already working")>]
 let ``Introspection executes an introspection query`` () =
-    let root = Define.Object("QueryRoot", [
+    let root = Define.Object("QueryRoot", [|
         Define.Field("onlyField", String)
-    ])
+    |])
     let schema = Schema(root)
     let (Object raw) = root
     let result = sync <| schema.AsyncExecute(parse Introspection.introspectionQuery, raw)