DO NOT MERGE YET

This is a bunch of possible optimizations including:

- "Compilation phase" - before, the field execution was resolving field values based on their type while running a query. After this PR, all steps made after field value resolution can be attached ahead of time, during schema creation. /cc #32
- Removed `Choice<obj,exn>` which was returned for each field to inform if resolution was performed successfully or if there is an error to log. Right now error logging is direct.
- Moved from F# lists to arrays - since most of operations are iterations, this should be positive effect too.
- Changed field collection phase - previously it was using a lot of list traversals, function calls and results aggregation. Right now it's more (ugly) iterative process with loops and mutable collections - but it's also faster.

Unfortunately, due to issue https://github.com/PerfDotNet/BenchmarkDotNet/issues/214 , I'm not able to make reliable benchmarks of the current solution.

Future work:
- Optimize synchronous paths - right now we're using async as uniform way to resolve any kind of field, no matter if it's synchronous or not. Great optimization here would be to distinct between sync/async code in compilation phase and generate field executors depending on each of those cases.
-  Use [Hopac](https://github.com/Hopac/Hopac) - Async computation expression is known of it's bad performance. And since it's used quite heavily (because of the point above), Hopac could potentially help here a lot.