@Horusiath 2016-06-20
First results after applied changes:

### Before

```ini

BenchmarkDotNet=v0.9.7.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU 3.00GHz, ProcessorCount=4
Frequency=2929687 ticks, Resolution=341.3334 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=SimpleExecutionBenchmark  Mode=Throughput  

```
                       Method |      Median |     StdDev |        Mean |         Min |         Max |     Op/s |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
----------------------------- |------------ |----------- |------------ |------------ |------------ |--------- |------- |------ |------ |------------------- |
 BenchmarkSimpleQueryUnparsed | 105.1598 us |  1.4777 us | 105.1218 us | 102.2005 us | 107.0542 us |  9512.77 |  48,06 |     - |     - |           7 236,65 |
   BenchmarkSimpleQueryParsed |  55.8036 us |  0.9719 us |  55.9251 us |  54.1766 us |  59.3268 us | 17881.07 |  26,14 |     - |     - |           3 914,84 |
   BenchmarkFlatQueryUnparsed | 249.1568 us |  3.7191 us | 249.8558 us | 245.2209 us | 258.3915 us |  4002.31 |  96,12 |     - |     - |          16 001,75 |
     BenchmarkFlatQueryParsed | 174.5001 us |  1.8256 us | 174.9276 us | 172.5124 us | 180.1565 us |  5716.65 |  17,66 | 12,41 |     - |           8 340,25 |
 BenchmarkNestedQueryUnparsed | 519.4034 us | 11.6238 us | 523.4641 us | 508.9341 us | 543.5234 us |  1910.35 | 578,00 |     - |     - |         104 652,58 |
   BenchmarkNestedQueryParsed | 287.8704 us |  4.7698 us | 289.6482 us | 283.6765 us | 305.2006 us |  3452.46 | 357,92 |  5,25 |     - |          72 755,49 |

### After

```ini

BenchmarkDotNet=v0.9.7.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU 3.00GHz, ProcessorCount=4
Frequency=2929687 ticks, Resolution=341.3334 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=SimpleExecutionBenchmark  Mode=Throughput  

```
                       Method |      Median |    StdDev |        Mean |         Min |         Max |    Op/s |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
----------------------------- |------------ |---------- |------------ |------------ |------------ |-------- |------- |------ |------ |------------------- |
 BenchmarkSimpleQueryUnparsed | 329.4432 us | 2.3003 us | 329.3496 us | 324.6654 us | 332.9564 us | 3036.29 |  38,61 |     - |     - |           6 221,65 |
   BenchmarkSimpleQueryParsed | 274.8102 us | 3.8274 us | 273.9460 us | 265.6387 us | 281.8587 us | 3650.35 |   8,89 |     - |     - |           2 178,75 |
   BenchmarkFlatQueryUnparsed | 354.8151 us | 2.6875 us | 354.2525 us | 349.1364 us | 360.3077 us | 2822.85 |  64,99 |     - |     - |           9 648,61 |
     BenchmarkFlatQueryParsed | 273.6142 us | 3.2951 us | 274.5027 us | 268.2144 us | 281.3114 us | 3642.95 |   9,09 |     - |     - |           2 223,63 |
 BenchmarkNestedQueryUnparsed | 509.8961 us | 4.5644 us | 511.0484 us | 503.1488 us | 518.8788 us | 1956.76 | 226,00 |     - |     - |          34 556,83 |
   BenchmarkNestedQueryParsed | 277.2645 us | 2.5692 us | 276.5373 us | 270.1634 us | 279.8317 us | 3616.15 |   9,09 |     - |     - |           2 217,65 |

### Conclusions 

1. Don't trust benchmarks :) 
2. It looks like the new implementation has more predictable speed, as it's similar for both simple and complex queries.
3. For simple queries new implementation is definitely slower than the previous one. For complex queries it's a little faster.
4. In terms of object allocations, new implementation is definitely better. A lot less heap allocations, and in all cases none of them is promoted over Gen0. Also you can note that over 80% of all allocations in new impl are done by query parser.

### Next steps

I'm quite surprised with such weak performance improvement (in case of complex queries). Need to investigate it further.